Feat/4-02/adding products to cart auth anonymous

+++ src/app/features/cart/cart-page/cart-page.component.ts
@@ -1,4 +1,4 @@
-import { Component, inject, OnInit, signal, WritableSignal } from '@angular/core';
+import { Component, inject, OnInit } from '@angular/core';
 import { ButtonComponent } from '../../../shared/ui/button/button.component';
 import { NavigateToSpecificRouteService } from '../../../shared/services/navigate-to-specific-route/navigate-to-specific-route.service';
 import { CartListItemComponent } from '../components/cart-list-item/cart-list-item.component';
@@ -6,8 +6,8 @@ import { RouterLink } from '@angular/router';
 import { CartService } from '../services/cart.service';
 import { CartResponse } from '../interfaces/cart-response';
 import { CurrencyPipe } from '@angular/common';
-import { CartModel } from '../model/cart-model';
 import { FormsModule } from '@angular/forms';
+import { MatSnackBar } from '@angular/material/snack-bar';
 
 @Component({
   selector: 'app-cart-page',
@@ -16,66 +16,63 @@ import { FormsModule } from '@angular/forms';
   styleUrl: './cart-page.component.scss',
 })
 export class CartPageComponent implements OnInit {
-  private cartId = '88c91569-25d0-4291-9738-3ad7c36d1c80';
-
+  public navigateToSpecificRouteService = inject(NavigateToSpecificRouteService);
+  public cartService = inject(CartService);
   public showCode: boolean = false;
   public showCodeMassage: boolean = false;
   public isCodeSuccess: boolean = false;
-  public navigateToSpecificRouteService = inject(NavigateToSpecificRouteService);
-  private cartService = inject(CartService);
-
-  public cart: WritableSignal<CartModel | null> = signal<CartModel | null>(null);
   public couponCode: string = '';
 
+  private _snackBar = inject(MatSnackBar);
+
   public ngOnInit(): void {
-    this.cartService.getCartById(this.cartId).subscribe((response: CartResponse) => {
-      this.updateCartModel(response);
+    this.cartService.handleCart();
+  }
+
+  public openSnackBar(message: string, action: string) {
+    this._snackBar.open(message, action, {
+      duration: 5000,
+      horizontalPosition: 'right',
+      verticalPosition: 'bottom',
     });
   }
 
-  private updateCartModel(response: CartResponse): void {
-    this.cart.set(
-      new CartModel(
-        response.version,
-        response.id,
-        response.lineItems,
-        response.totalPrice,
-        response.totalLineItemQuantity,
-        response.discountOnTotalPrice?.discountedAmount.centAmount,
-        false,
-      ),
-    );
-  }
-
   private getSelectedCartItem(productId: string | undefined) {
-    return this.cart()?.lineItems?.find((item) => item.productId === productId);
+    return this.cartService.cart()?.lineItems?.find((item) => item.productId === productId);
   }
 
   public increment($event: string | undefined): void {
     const selectedItem = this.getSelectedCartItem($event);
     const quantity: number | undefined = (selectedItem?.quantity ?? 0) + 1;
 
-    this.cartService.changeLineItemQuantity(this.cart()?.id, this.cart()?.version, selectedItem?.id, quantity).subscribe((response) => {
-      this.updateCartModel(response);
-    });
+    this.cartService
+      .changeLineItemQuantity(this.cartService.cart()?.id, this.cartService.cart()?.version, selectedItem?.id, quantity)
+      .subscribe((response) => {
+        this.cartService.updateCartModel(response);
+      });
   }
 
   public decrement($event: string | undefined): void {
     const selectedItem = this.getSelectedCartItem($event);
     const quantity: number | undefined = (selectedItem?.quantity ?? 0) - 1;
 
-    this.cartService.changeLineItemQuantity(this.cart()?.id, this.cart()?.version, selectedItem?.id, quantity).subscribe((response) => {
-      this.updateCartModel(response);
-    });
+    this.cartService
+      .changeLineItemQuantity(this.cartService.cart()?.id, this.cartService.cart()?.version, selectedItem?.id, quantity)
+      .subscribe((response) => {
+        this.cartService.updateCartModel(response);
+      });
   }
 
   public delete($event: string | undefined): void {
     const selectedItem = this.getSelectedCartItem($event);
     const quantity = 0;
 
-    this.cartService.changeLineItemQuantity(this.cart()?.id, this.cart()?.version, selectedItem?.id, quantity).subscribe((response) => {
-      this.updateCartModel(response);
-    });
+    this.cartService
+      .changeLineItemQuantity(this.cartService.cart()?.id, this.cartService.cart()?.version, selectedItem?.id, quantity)
+      .subscribe((response) => {
+        this.openSnackBar('Item has been deleted successfully.', 'Close');
+        this.cartService.updateCartModel(response);
+      });
   }
 
   public buttonGoToCatalog(): void {
@@ -103,11 +100,11 @@ export class CartPageComponent implements OnInit {
     this.isCodeSuccess = true;
     this.handleCouponCodeErrorMessage();
 
-    this.cartService.addDiscountCode(this.cart()?.id, this.cart()?.version, this.couponCode).subscribe({
+    this.cartService.addDiscountCode(this.cartService.cart()?.id, this.cartService.cart()?.version, this.couponCode).subscribe({
       next: (response: CartResponse) => {
         this.isCodeSuccess = true;
         this.handleCouponCodeErrorMessage();
-        this.updateCartModel(response);
+        this.cartService.updateCartModel(response);
       },
       error: (error) => {
         console.error('[cart apply code error]', error);
@@ -118,16 +115,20 @@ export class CartPageComponent implements OnInit {
   }
 
   public clearShoppingCart(cartId: string | undefined, cartVersion: number | undefined): void {
+    const confirmation = confirm('Are you sure you want to empty this cart?');
+
     if (cartId) {
-      this.cartService.deleteCartById(cartId, cartVersion).subscribe({
-        next: () => {
-          this.cart.set(null);
-          this.cartService.cartItemsCount.set(0);
-        },
-        error: (error) => {
-          console.error(error);
-        },
-      });
+      if (confirmation) {
+        this.cartService.deleteCartById(cartId, cartVersion).subscribe({
+          next: () => {
+            this.cartService.updateCartModel(null);
+            this.openSnackBar('The basket has been cleaned successfully.', 'Close');
+          },
+          error: (error) => {
+            console.error(error);
+          },
+        });
+      }
     }
   }
 }
